NDEV-3498 upgrade test
